Move models outside entities
