+++ src/models/petsTypes.ts
@@ -1,5 +1,5 @@
 import { baseModel } from "./baseModel";
-import { generalOptions } from "../mongodb";
+import { generalOptions } from "../entities/mongodb";
 
 export interface RequestGetPetsTypes {
     key?: string;
