+++ src/services/auth.ts
@@ -1,5 +1,5 @@
 import { helpers, infrastructure } from "../app";
-import { RequestLogin } from '../entities/models/auth';
+import { RequestLogin } from '../models/auth';
 import bcrypt from "bcryptjs";
 
 export class AuthServices {
